cleaned code and complied with norm
